[fix] model job and cpn
