+++ express_service/services/job/configs/index.js
@@ -18,5 +18,5 @@ module.exports = {
   DB_MYSQL_HOST,
   RABBITMQ_CONNECTION_STRING,
   DB_PG_URI,
-  GRPC_JOB_SERVER
+  GRPC_JOB_SERVER,
 };
