+++ express_service/services/job/models/index.js
@@ -4,11 +4,9 @@ const Company = require('./company');
 async function syncModals() {
   Company.hasMany(Job, { onDelete: 'CASCADE' });
   Job.belongsTo(Company, { onDelete: 'CASCADE' });
-  
-  /* await Company.sync({ alter: true });
-  await Job.sync({ alter: true });
-  await Product.sync({ alter: true }); */
 
+  /* await Company.sync({ alter: true });
+  await Job.sync({ alter: true }); */
 }
 
 module.exports = syncModals;
